summary: build_a_s_query always returns a query
There was a test that was supposed to verify that
build_alert_summary_query always returned a query. The test, however,
didn't use proper permission flags, so it ended up testing nothing, and
the code, meanwhile, started returning false for users which aren't
authorized for anything (if they aren't, they shouldn't have been
allowed access into ninja, so it's a bit of a moot point, but I don't
want to delete tests discovering a violated invariant, even when
said invariant is only violated in a special case that shouldn't be
reproducible - the invariant should still hold).

Most importantly, the test that is changed to show the problem makes
ninja build again after recent changes in op5lib.

Change-Id: Ib95d694d0b7789ee948dad9c554d8a155bd779b6
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>