Remove date helper from Kohana
The date helper in Kohana implements a couple of methods regarding
dates, such as creating an array containing 00-59, or subtracting 12 if
another string is 'pm'... or converting a time interval to an array
with hours, minutes and so on, in a strange way.

Nothing depends on that helper, but it's creativly extended in a date
helper in application/helpers/MY_date.php.

Remove the date class from kohana, and make the date class in
application standardized, so we can skip the creative class-mangling we
have.

Change-Id: I1e6e04505f14c7fd4f023a910cc1f0e0827a6b1a
Signed-off-by: Max Sikstrom <msikstrom@op5.com>