cookie: Separate 'secure cookie' and login logic
Through the cookie.secure configuration option, we allow admins to
choose to only accept logins through HTTPS, which is also the default
option. This causes confusion, since it really regards "logging in" and
not "persisting information over multiple HTTP requests".

To clarify, we want to send an error message whenever a user wants to
login through HTTP but Ninja's setting is to only allow HTTPS. We use
the session which is configured to set a secure cookie, thus we can
never read it back from plain HTTP. This solution always allows
arbitrary cookies to be set, based on 1) availability of HTTPS; but
falls back to 2) the configuration value. This is complemented with
another check in the login process, which demands coherence between the
current protocol being used and the one that is configured.

Previously, we tried to let the user login through HTTP but never
notified her about the HTTPS default setting, and she was redirected to
the login form once again.

This fixes #7595

Change-Id: Iad94f3b11678e4692f64969d5bda1aa6c2f1b9f4
Signed-off-by: Carl Helmertz <chelmertz@op5.com>