Remove calculator
We had it for a while, it was a nice test, and it fulfilled it's
purpose of beeing a playground for grammar and parser generators, at
the same time as beeing a little easter egg.

So it's time to remove it, strip a second of the ninja build time
(probably not more at least), and remove a couple of lines of code.

Bye calculator, you won't be missed.

Change-Id: I051f317a3ca65f79970960edca71cd3edb611c5f
Signed-off-by: Max Sikstrom <msikstrom@op5.com>