From hostgroup details, request groups explicitly, not by in-operator
When showing hostgroup summary, it's a list of hostgroups, not object
sets in host domain. Klicking on a host group name would therefore be
expected to show the host group itself, not the named set with the same
name, that might be the hostgroup. Therefore use [hosts] groups >=
"name"-filter instead of [hosts] in "name"

Same applies to service groups

Change-Id: Ie0e5e04e02749842e8f526a376131255a0efe9cd
Signed-off-by: Max Sikstrom <msikstrom@op5.com>