listview: List commands
I don't see any reason to withhold the awesomeness of listviews from
commands. IMO, this is coherent to the rest of the product. We also take
away one more incentive for "View config" to exist.

This patch exposes the 'commands' table by adding an <option> to the
graphical filter builder, no further menu items are made visible.

Change-Id: Ib5dcd3698ee9ee04016bd889a843ea059828a6c6
Signed-off-by: Carl Helmertz <chelmertz@op5.com>