backup: Run backup/restore as monitor and fix up file permissions
Since these programs are no longer set-UID-root, we need to run them as
the user which is actually supposed to be able to read/write the files
which we backup.

This patch also ensures that permissions are preserved for auth config
backups.

Since cli/upgrade_auth may be, and often is, run by root during a %post,
we want to preserve ownership and mode for the configuration in order to
not render them useless.

I also, by recommendation from HS, changed the name of the resulting
backups to be suffixed with the timestamp, instead of being infixed
with it.

This fixes #8279.

Change-Id: I15af7a5103aaab4122f5bb7d9d186806dc41beca
Signed-off-by: Anton Lofgren <alofgren@op5.com>