cli/upgrade_auth: Preserve permissions for auth config backups
Since this CLI may be, and often is, run by root during a %post, we
want to preserve ownership and mode for the configuration in order to
not render them useless.

I also, by recommendation from HS, changed the name of the resulting
backups to be suffixed with the timestamp, instead of being infixed
with it.

This further fixes #8279.

Change-Id: I15af7a5103aaab4122f5bb7d9d186806dc41beca
Signed-off-by: Anton Lofgren <alofgren@op5.com>