Replace config_replace_map with specific methods in Auth
All keys in configuration is hard coded in the routins using the
configuration. Management of objects related to configuration (like
auth) is accessed within each modules/library. Also keep updates in
those libraries.

Doing so, it's much easier to change the algorithm for renaming later,
if nessecary. For example, if we want to add group renaming so it
queries LDAP or something similar

(This commit is imported from op5lib)

Change-Id: Ib517b5295d438c42e48a4d14f60ad7dd1ac0e84e
Signed-off-by: Max Sikstrom <msikstrom@op5.com>