auth: Check for modules before trying to verify membership
A user needs the modules config key set in auth_users.yml. It can be
empty though. The reason behind doing these checks in the driver class is we
have no way of checking the users defined modules with the name of the module
using the Default driver at an earlier stage of the authentication process.
To ease the pain we try to bail as early as possibleif the user is not allowed
to use this particular module

(This commit is imported from op5lib)

Change-Id: Ic8884c4363345606bc5f45156cf145d90910fdd3
Signed-off-by: Carl Helmertz <chelmertz@op5.com>
Signed-off-by: Henrik Smederod <hsmederod@op5.com>