ninja_widget: Fallback through user's tac/index
This means that ninja will:
 * Remember your settings, per-page
 * Fallback to the old widgets, per-page, so when you clone a widget, it
   will really be cloned.
 * Fallback to any (manually) configured user-specific defaults,
   per-page.
 * Fallback to any manually configured page-specific default for
   all users (like creating different thresholds for a
   simplified/complexified tac, say).
 * Fallback to the user's tac/index (meaning that by default when you
   open tac/myother, you get the nagvis maps etc that you've previously
   selected for tac/index, and get to customize it from that starting
   point, rather than the global default)
 * Make anything else inherit from the global default.

This commit inserts the second-to-last point in the list - the rest is
old news.

This only really applies to widget settings - widget orders are handled
differently.

This fixes bug #6314, however, note that when you first open the
external widget, you can only continue to configure it through the
widget, not through ninja's tac/index - partly out of laziness, true,
but it effectively ensures that I don't break anything for anyone who
worked around that bug.

It has been suggested that external widgets should be impossible to
configure through the external widget interface, but that's a separate
bug - in fact, it's two: #5525 and #7236. This commit does not fix that.

And I think this is a new personal best for the
lines-of-commit-message-to-lines-of-patch ratio at 37:1!

Change-Id: If077747a8c6fd2ff7410570c2d5c642c45ad5d80
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>