config: Replace Livestatus interface with ORM
                                DADE
                    It has a killer refresh rate.

                                KATE
                    P6 chip. Triple the speed of the SQL.

                                DADE
                    Yeah. It's not just the chip, it has
                    multiple backends. But you knew that.

                                KATE
                    Indeed. ORM architecture is gonna change
                    everything.

                                DADE
                    Yeah. ORM is good.

This also gets rid of the old config model - while it had more detailed
support for timeperiods that is now missing, we haven't actually
rendered the daily ranges for a while, and nobody noticed - in fact, we
didn't render anything at all, so apparently nobody uses this page
anymore. Either way, the "in effect?" column is probably usually more
helpful, as the old version didn't dump exception or exclusions.

This also starts to render commands again.

This also starts to render hostgroups again, fixing #8139.

Change-Id: I8b897d043cfcbd4b3d4a8131b9074ec242429b7f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>