Use _expanded-columns in livestatus instead of own macro expansion
But we still want to have the name notes_url, action_url and notes, but
we want to used the notes_url_expanded and so on in livestatus. That
means we need to use the new rename-api in the ORM builder.

This depends on the change Ia9b8c78cf51be262daa9972b2d47cf9676a6bc9a in
op5lib:
Make it possible to use different columns in backend and ORM (rename)

Change-Id: I76e7fbc4714c7ad02150680601c897f8f717ea88
Signed-off-by: Max Sikstrom <msikstrom@op5.com>