Throw out macro resolution in ORM
only config_url still uses macros, but only the name-related macros is
used. Simply hard code that column in the hosts and services table,
where it's used.

Also add tests that config_url contains the hostname for hosts, and
that the host name isn't empty, which is a minimalistic test that
config_url still works with macro resolution.

Also fix a minor perfdata related test at the same time

Change-Id: I1c5cff577ba3f5fe5d8f519a3ba593d014852bdd
Signed-off-by: Max Sikstrom <msikstrom@op5.com>