ORM: Remove the explicit layer of LS/SQL abstraction
For set:s, we had the following layers of abstraction (in this example,
services):
- ServiceSet - wrapper class, for application specific logic

- BaseServiceSet - generated, with per-object-mappings, like column
name translations

- ObjectLSSet - wrapper class for generic applicaiton specific logic on
top of livestatus wrapper

- BaseObjectLSSet - Access methods for livestatus

- ObjectSet - wrapper class for applications specific logic on generic
sets

- BaseObjectSet - Contains set operations, generated class

In fact, we never use the ObjectLSSet, and in BaseObjectLSSet, we need
to know about field names and structure, so we need to move out those
things to BaseServiceSet instead.

So refactor, remove ObjectLSSet and treat this inheritance as
inheritance in the generators instead, which means the code for
accessing the backend will be available in each type-class instead.
That means, it can be adapted with generated vaildators and
column-renaming.

This patch itself won't solve the bug with JOIN:s in SQL not resolving
the names correctly, but it will build the foundation about solving it.

This is refactoring, now we set the tests on trial, and see that the
refactoring doesn't break anything.

(This commit is imported from op5lib)

Change-Id: I7cbdbbc093057eea8bf5d8e4fd66011272809345
Signed-off-by: Max Sikstrom <msikstrom@op5.com>