ORM: Refactor process_field_name to do validation
Moving more validation to process_field_name (and rename it to match
what it does better) elimintes the need of a specific method to
validate and filter out columns, and another to rename columns.

Update process_field_name to map_name_to_backend, which takes one
argument, a ORM field name, and returns the corresponding name of the
backend field, including relationship resolution, or false if columns
isn't available.

Using that method everywhere, vaildation if the column exist, is valid,
and safe, can be done much easier.

This also changes the API for LivestatusFilterBuilderVisitor, so we
send a callback to that method instead. This simplifies testing, and
makes the API match what earlier was used in SQL.

This fixes the bug of name resolution in SQL backend, if the correct
patches is applied to ninja too. all modules providing tables needs to
be rebuilt after this patch is applied.

(This commit is imported from op5lib)

Change-Id: I1cea539cd09fdcb5b4221e50e9af6fc51c1cac92
Signed-off-by: Max Sikstrom <msikstrom@op5.com>