log: Let log files be group writable
Since the op5Log class is not always invoked by the same user, it is
necssary to allow members (like monitor) of the same group (usually
apache) as the owner (usually apache) to write to logs as well.

This fixes #8540.

This also fixes a lurking bug in the chown/chgrp call a few lines down,
which would render the log unwritable by anyone but the monitor user
(read: not writable by apache) if the root user ever attempted to log
something.

(This commit is imported from op5lib)

Change-Id: I2359b6e6c39a7fe0847967d682882209d296157d
Signed-off-by: Anton Lofgren <alofgren@op5.com>