recurringdowntime: Only unserialize strings
When the object comes from the web form, it'll be full of arrays, not
serialized strings.

It will come from the web form when objects aren't selected, which will
lead to the form trying to render a validation error, but crashing due
to the unexpected format of this parameter.

This resolves bug #8588.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>