livestatus: Stop relying on broken short-cut
Earlier versions of the livestatus library had a short-cut, where you
could give a string as a column name, in which case the library would
assume you wanted less deeply nested arrays back - which as if by
accident happened to correspond to what we need whenever we want to get
a list of names.

This version of the livestatus library's support for stat responses
seemingly makes the shortcut appear to work, except the query building
code doesn't support the shortcut, so that instead of restricting the
columns in the query to what we'll be returning, we download ALL THE
DATAS! This has the downside of large installations eating up tons of
memory and crashing horribly whenever a stupidly written form is about
to be rendered.

This works around that "quirk" in a minimally invasive way, by no longer
relying on that shortcut in all the places I've found it to be used.

The PNP helper by design throws an exception, so it won't run in a
browser, but it showed up in my grep. The execute_command code
(bug #8617) should no longer be trivially triggerable, but was manually
verified when the same change was made for 6.2. The tests passes.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>