cucumber: Update helptext step
The upgraded qtip no longer reacts to click events, but it does react to
mouseover.

However, it only does so with the help of jquery animations, so we must
also sleep, and it would appear ruby's sleep only takes integers.

To prevent similar bugs from going undetected, we also enforce non-empty
helptexts.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>