reports: Move helptexts away from summary
Because avail and SLA reports in op5reports needs summary reports'
helptexts to be defined, just move them so they are.

We could have just implemented a fallback in op5reports, but since the
"404" condition is a random string, it would have required us to regex
when to fallback, and I really don't want to.

This fixes bug #7831.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>