Make it possible to sort on simple custom columns
Sorting on custom columns is in general impossible, since the backend
that sorts on the columns isn't aware it's possible to sort on.

But in some specific cases it is. And for now, those cases are if the
column only contains a variable that exist in the backend and that
variable is of type string, int or time.

Resolves: #8158

Change-Id: Ieeb49fa70c71d04381e0e80ad6dc6bfe0001d907
Signed-off-by: Max Sikstrom <msikstrom@op5.com>