reports: Remember selected values when using empty group
This fixes bug #7836, by simply urlencoding the parameters.

We've kept saying that we can't urlencode the parameters, because they
might not fit, but there's a logical upper bound to how many empty
groups it's reasonable to select, so I don't worry too much.

Yes, this does go straight for the inner array of the report option
array, which it shouldn't. However, it's the one time I don't want to
urlencode any ampersands, because I'm not using the URL in a HTML
context.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>