lsfilter: Stop frickin' reloadin' all' the' time'
This fixes #8533, not only for textareas but for the visual editor as
well by refactoring the lsfilter_main object.

Now the lsfilter_main object emits signals to each listener, inviting
everybody who wants to, to join the "receive updates about the
listview's global state" club.

update_run() is now documented to as a private method, you should always
trigger updates through:
- update() // it takes args
- update_delayed() // same as update() but not as eager
- refresh() // triggers update(), reusing the arguments from last time

The central try-catch construct in update_run() is now limited to a
single line which makes it much clearer where the method's exit points
are.

Change-Id: I40b1d0909d8c3cd6588a2e96a9c7dcd3b266625f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>