reports: Load custom SLA timeranges lazily
We don't need to *always* check for the oldest report event in the
database - most of the time, the user isn't generating a custom SLA report.

Because a SLA form is a special case of the avail form, I can't quite
detect the difference between a custom SLA report and a custom avail
report, so it will be loaded on both. It's a light query, though, so I
don't particularly care.

Change-Id: I5a2d0282bf922522da5cb2b5ef7d02304fa21619
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>