ORM: Rename timeperiod field "in" to "is_active"
since, "in" is a reservced keyword in the filter language, we can't use
it as a column name for filtering, and thus none could ever have used
it.

"[timeperiods] in = 1" would end up with a parse error at =, since it
expects a saved filter name there, not an operator.

The only way this coulumn would be possible to use is in the new filter
api, which is still possible since the rewrite method in the timeperiod
model.

Change-Id: Id766bbf54caeee701c04f1d17c13ea9a6c29b6ec
Signed-off-by: Max Sikstrom <msikstrom@op5.com>