Rewrite how reports are stored
Instead of storing the reports in tables that are unique to each report
type (some of which require insane amounts of migrations, while others
store report data serialized which breaks), let's keep the report
options schemalessly as key-val pairs in one table, associated to the
report by its ID.

To do so, we teach the Report option structure to persist and restore
itself - while it's quite bloated, it has nice properties that I want:
it's fairly strictly typed, while being easily extendable on a
per-report-type level. This resolves bug #5624.

As a side-effect of this typing and extendability, this commit also
resolves bug #7897.

Yes, it turns out I was writing another ORM layer all along. But I
didn't know it at the time!

Change-Id: Ic61d1651a142a8b399455377050936d044b757db
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>