histogram: Make histograms crash less
They used to crash whenever you tried to view a host histogram for a
host that had an unreachable service.

This fixes bug #7751. It could be that it's still crash:able, but my
brand new, relatively encompassing suite of tests should be easy to
extend if new ways of achieving that are discovered.

As described in comments for #7751, there is a fundamental problem with
how our histograms are presented - therefore, this moves histograms to
the main menu, and because I literally got saving for free, I'm throwing
that in as well, at no extra charge.

Finally, this moves the histogram views to the report module where they
belong, and some unused code - such as a custom CSS with no relevant
content - are deleted.

Change-Id: I4955df5b16fbccf1ace7a93587eaab8191026bdb
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>