lsfilter: Use server's timezone
Because everything else uses server time (or, to be precise, PHP time,
which is totally unrelated to the actual server time, but I'm not even
going to bother), rather than client time, the easiest way to make
listviews consistent with everything else is to convert the client time
to server time. It would probably be more Correct™ to move in the other
direction, but it would take a *lot* more work.

This has been manually tested by putting my VM ('s PHP time) into US
Pacific time, and noticing that listviews, like default values in
command submissions and extinfo content, moves 9 hours away from my main
computer's CET timezone. Time has a habit of moving about, so I don't
quite know how to auto-test it.

So, despite the lack of tests, I'm going to say that this resolves
bug #7669.

Change-Id: I409869f92f5e4c01f4fe29e541eb9a6a46b67208
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>