httpapievent: Set in_scheduled_downtime to null always
After months of discussion about it, this is the final call for
bug #7804. No, this does not in any way help - this is just a way of
saying "wontfix - see bug #7968".

Setting it to null, rather than removing it, prevents any breakage.

I would document it as deprecated, except I can't, because bug #7984.

Change-Id: I203c01c7eab1c09560b7bae46fce5bfbc5324b3a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>