alert_history: Expand-Contract long output
Whenever there is long output present for an alert a control to expand
or contract the output is visible. By default the long output is
expanded if the "Include long output" option is checked and contracted
if the option is not checked.

Additionally the controls are hidden when the report is printed.

This implements #8417

Change-Id: Id44561658d0df9fc2f93892d334fdad16e45ddd7
Signed-off-by: Henrik Smederod <hsmederod@op5.com>