alert_history: Trim long output before outputing controls
If the plugin output consists of one line and ends in a newline the
long output would only contain a \n resulting in unnecessary
expand/contract controls that would reveal ... *drumroll* ... NOTHING!
Trim the long output and check it's length before outputing the
controls.

Also squeezed in a css class rename to shorten the classname.

Additional tweak for #8417

Change-Id: I9cc86a9aba310f163842ad9c5f63fc6b610fbf3a
Signed-off-by: Henrik Smederod <hsmederod@op5.com>