ListView: Don't load [hosts] all before the actual query
We used [hosts] all ass fallback for invalid queries in a strange way
earlier. Since the list views is refactored, this is totally incorrect,
making each page load first load [hosts] all before the actual query,
makeing it a race, which failed the tests once every 20 times or so...

Removing this should fix the racy builds.

Change-Id: I82cd35ca7305bbcb74173b674b0f16d7d6ab0780
Signed-off-by: Max Sikstrom <msikstrom@op5.com>