recurring_downtime: Reenable javascript validation
A number of commits caused the validation not to run. Not that it worked
after the changes anyway but hey let's make it a bit harder for the user
to make mistakes shall we.

All mandatory fields are now properly marked as such, all timestrings
are now validated instead of just half of them and the triggered by
validation is removed since it's was previously removed from the form in
cbe4efa.

While we are at it let's add a couple of error strings for all time
fields stating that :ss is optional.

Change-Id: If7c987a5aea82915cf972e113ab0dcdee24340c8
Signed-off-by: Henrik Smederod <hsmederod@op5.com>