Recurring downtime same day
When a new recurring downtime schedule is created today is checked
against the schedule and if today could be included the user is asked
whether a downtime should be inserted for today.
This is only true if a new schedule is created and not if a schedule is
updated.

This implements #4613

Change-Id: Idd37d0e61a04ed3f0b7eee4c3313726715e3b1f7
Signed-off-by: Henrik Smederod <hsmederod@op5.com>