schedule downtime: Handle wrong timestamp format more gracefully
The error handling was already in place, but not checking the return
value of DateTime::createFromFormat() made us try to execute
::getTimestamp() on false => fatal error.

This fixes #7272, not by allowing timestamps without seconds, but by
properly handling the error. The helptext already contains the full
format and adding :00 if the seconds are missing from the input, does
not comply with all timestamp formats one could possibly use (see
928aa3214a1e09865f244199cd7281a8502b6f84 which both introduced this
crash, but also argues for not resolving dates with string functions).

Change-Id: Ica34bd31ed092d719687f34076a068656674ce16
Signed-off-by: Carl Helmertz <chelmertz@op5.com>