Fix postgres database backend
By some reason the Kohana developers thought it would be nice to just
make the postgres driver half way finished, and just invent how some
methods should be used. Instead of just using it like the developer
wanted, use it as it is intended instead.

pg_field_name takes two arguments: A result set and a field index. It
doesn't take a result set and iterate over the fields. It would be a
working way of doing it however, IFF the API was designed that way.

Change-Id: I7abe32b6429362844566ce91fedf10786b15fd42
Signed-off-by: Max Sikstrom <msikstrom@op5.com>