lsfilter: Add support for non-timestamp date formats
This resolves bug #8572, though with a couple of caveats.

It introduces a date() "function" in the filter API, instead of merely
parsing a string as the bug suggested. This approach was suggested by
msikstrom, to be able to generalize the language in the future without
time columns' implicit magic being in our way.

While this introduces a "generic" function parser in the grammar, the
javascript side is implemented somewhat hacky - though unlike the PHP
side of things, we won't need to support the exact javascript behavior
forever, so that can be improved when the second function is
implemented.

The solution is provided by strtotime, because:
 * We already use that in a couple of places. Not that many, but it's
   not a brand new solution.
 * I didn't have to implement a date parser, which means I almost
   certainly didn't create any bugs in the date parser.
 * It's well-documented:
   http://www.php.net/manual/en/datetime.formats.php contains a ton of
   documentation, while many others comes with a supposedly
   confidence-inducing sentence that claims it understands "most" dates.
 * It's got loads of nice and readable relative date representations -
   while it doesn't understand when it was Katy Perry danced on table
   tops, took too many shots, and might or might not have kissed (in
   it's defence, that probably took more than a second, so..), it *can*
   parse "last friday" or "last friday midnight", as well as "four hours
   ago". In comparison, at's insistence that "4pm + 3 days" is all you
   need seems positively quaint. Of course, if you like it, strtotime
   supports at's syntax as well.

This is also bad, as it won't (ever) be parsed correctly by javascript
or anything else - a full parser would have been better, and would have
allowed more flexibility and interactivity in the future. However, I'm
assuming that relative date strings is what this will be most used for,
and I don't think a full javascript parser would have been able to do
much fun with those, anyway - for instance, a date/timepicker would be
actively harmful, as it would distract from this purpose. Also, the fact
that it's a function rather than a built in part of the language makes
the "PHP only" part of it more palatable.

This does not care about the user's configured date format at all,
because doing so would mean any date-related saved filter would break
when you change how dates are represented. If we redefine nagios'
date_format key for euro so it uses "DD.MM.YYYY" instead of
"DD/MM/YYYY", it would not only be more euro-friendly(!), but also
different-looking enough from the US option, "MM/DD/YYYY", for strtotime
to do the right thing in both cases.

The query builder automatically unwraps/rewraps the date, so you can't
type timestamps into the query builder anymore, yet you can write "5
minutes ago" or "midnight yesterday" and have them be parsed for you
automatically.

tl;dr:
[services] host.last_state_change > date("midnight yesterday") and
           last_state_change < date("5 minutes ago")

Change-Id: I4e3635ac86d5c50339e420e704e9baed0bd36c69
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>