recurringdowntime: Let limited users see their recurring downtimes
The limited authorization code accidentally based its decisions on the
full pool, instead of the empty pool, which means that the number of
objects in the downtime the user is allowed to see, plus all the objects
the user is allowed to see, had to match the actual number of objects in
the downtime.

Hence, it was possible for limited to be allowed to see schedules, but
only if they weren't supposed to see them.

This resolves bug #8807.

Change-Id: Ibdc96d97a7a5fa77d290ddd03a8e4b36bfbfcdb5
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>