Add PostgreSQL visitor for ORM, again..
There are some incompatibilities with mysql, such as regexp searching.

The visitor itself is half boilerplate to identify the different
operators and nodes, and the other half is just configuration for the
syntax, and it's interleaved. Using the same class would just be
messier than treating it as configuration and copy everything. Also,
fixing bugs in mysql would possibly end up with creating bugs in
postgre if not copied

This is part of the solution of fixing regexp searching in
logserver-table

By some reason, this visitor become removed when refactoring from 7.0
to 7.1, and also the reference to the visitor got in the wrong way.

Without this patch, logger won't have regexp support.

Change-Id: I24223b05139c7a55e7feeaae7f451f51d80fbe5e
Signed-off-by: Max Sikstrom <msikstrom@op5.com>