Add PostgreSQL visitor for ORM
There are some incompatibilities with mysql, such as regexp searching.

The visitor itself is half boilerplate to identify the different
operators and nodes, and the other half is just configuration for the
syntax, and it's interleaved. Using the same class would just be
messier than treating it as configuration and copy everything. Also,
fixing bugs in mysql would possibly end up with creating bugs in
postgre if not copied

This is part of the solution of fixing regexp searching in
logserver-table

Change-Id: I578039709f7ed5a02e8396cdaf3fcab6393e88ec
Signed-off-by: Max Sikstrom <msikstrom@op5.com>