Remove support for widget upload
Since widget uploading is a way to upload arbitrary code to the system
and execute it without too much effort and privileges, this needs to go
away.

Uploading widgets is still possible, just unpack the widget in a
directory in the widgets directory of a ninja module (or in the
application/widgets directory), and call the ninja cli interface to
register it.

Resolves: #7504

Change-Id: Idac232014219ba29595bf85090212daabe2c3cba
Signed-off-by: Max Sikstrom <msikstrom@op5.com>