Kohana: Load include path at setup time
The include path was calculated sometimes before it's needed. We know
that we need it every page load, so we don't need the laziness. But if
we know that we always do it at setup time, we can mangle the path
afterwards, for example if we want to mock a module for external unit
tests.

Change-Id: I7e609576cff21e03d67c961e1fc25929e7708674
Signed-off-by: Max Sikstrom <msikstrom@op5.com>