Let views take responsibility for js, css
The guy responsible for making sure that the <style> and <script> tags
gets rendered according to what each controller wants is the view guy,
not the controller guy.

In the past, each controller had a bunch of copy-pasted code for
managing this individually, but that, well, sucked. I previously tried
to improve it by moving the responsibility to the ninja controller,
which looked like it worked, except for in PDF reports - PDF reports
aren't auto-rendered, so they don't trigger the auto-render hook in the
ninja controller that hooked into the view and transfered the resource
list.

This moves the responsibility further, to the view (in fact, to the base
template), which is where the responsibility belongs.

This fixes bug #8848.

This obviously breaks reports, geomap, and other modules that used the
old way. Patches for those will be coming up shortly.

Change-Id: I3a628065a3688ba477fb25b594fd59d065a82657
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>