Make sure template->css and template->js is arrays
Since many places (in this particular case the widget page) mangles
$this->template->css and $this->template->js as arrays, by appending
data to them, or worse, sorting them, without making sure they are
arrays, there might be a problem.

The case which made this problem appear was the widget page, which does
array_unique on the js array, but without adding javascripts to the
array, the array is unset, and a stack trace is generated.

By always initializing those to an empty array, we can ensure that they
are arrays and nothing else.

Change-Id: Ie3ee3701de9c1a086103c499bf9a933123417d93
Signed-off-by: Max Sikstrom <msikstrom@op5.com>