Auth: Make sure all user variables are defined after constructor
We depend on some fields in the user object to be defined, even if not
set. Like auth_data should be an array, realname needs to be readable
without an error/notice, so those are set when creating an User. But
the contructor replaced the entire storage. Merging in the new
parameters solves this issue, by keeping old values if not set.

(This commit is imported from op5lib)

Change-Id: Ie2bcdc32f57c5487b6acbd60cee5d25b6973382f
Signed-off-by: Max Sikstrom <msikstrom@op5.com>