Seperate listview/ORM from monitoring model/renderer
From the beginnning, the listview used to present the information in
the ORM layer representing the monitoring information from livestatus.
The natural separation was then to have ORM as a module interacting
with livestatus, and listview to build a frontend to the ORM layer.

The ORM has since then been used to much more, and been used in a much
more generic fashion. The orm and listview has been more tightly
coupled, and other parts have started using the ORM as interface. The
ORM has since then been used as a generic interface to access
information.

The monitoring part of the ORM has however been a secondary focus.
Since always, the monitoring information is the central part of the
product, but it is put at an equal level as log handling or trap
handling. And also, at an equal level as saved filters and soon saved
reports.

So split op the ORM module in a backend part, put in the lsfilter
module (soon to be renamed to listview), and the monitoring part to a
monitoring module.

Change-Id: Id63ffc0c070448660440c9856087487a30bd5e54
Signed-off-by: Max Sikstrom <msikstrom@op5.com>