Remove some unused libraries in Kohana
Since we have lot of old strange libraries in Kohana, and since long
have just used Kohana as a lightweight framework, we should remove the
objects we don't use and never will. If we need the functionality in
the future, we should just add libraries which does what it should
instead.

Change-Id: Ia3a48ca23c68806bceac1b62970e0852573f292f
Signed-off-by: Max Sikstrom <msikstrom@op5.com>