db_migrations: Remove dependency on op5reports
Because op5reports imported its own report information into the old
report tables, writing a migration into schemaless reports is tricky
without the op5reports schema. However, because we don't know at upgrade
time whether op5reports has been upgraded yet or not (actually, we can
be sort-of confident that it hasn't, because ninja comes before
op5reports in the alphabet), we don't know which version of op5reports
we would have to interact with.

Luckily, op5reports is nothing but a summary report and a PNP checkbox
inside an avail or sla report, and all the tricky cases for a report are
common between summary and avail/sla. So building a dumb-ass op5reports
reimplementation inside of the migration is feasible, and is therefore
what I did here.

This fixes bug #8912.

Testing this revealed that saved summary reports that applies to All
objects was busted in the save refactoring. This fixes that as well.

Change-Id: I4b2d08dfb96a273c84520d00377b0c5edd97f223
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>