Remove kohana cache write support
Since the internal cache can be written to a file, and since there are
no locks, and since there might be races, and since we never had tested
the cache with our patches, and mostly, since we never going to use it
as it is, we can most likely remove the possibility to write the
internal cache to disc.

If we need internal caching, and if we should add it again, we need to
check it's integrity anyway, and then we should probably replace it
with something like APC as backend, to keep the settings in memory
instead of wasting disc I/O more than nessecary.

Change-Id: Ib46e2e3e49a700944a102f20a5b628d7988893af
Signed-off-by: Max Sikstrom <msikstrom@op5.com>