nagstat::date_format: Don't be stupid
date_format returns a string that changes approximately never (I'd
estimate it at between 0-1 times per monitor deployment, with a strong
bias towards 0). It is retrieved by parsing all of nagios.cfg, and
looking for a key with an absolutely anemic set of options which we
really should implement in a way that doesn't suck so bad. On the other
hand, it's often retrieved many times per page load. So, it makes sense
to do some kind of caching, right?

Except... There are good ways of caching, and there are bad. And we
doubled up on the bad.

You could store the string in a static variable, and you'd only have to
do any work once per page load. It's simple, it's trivial, it's robust -
it's just common sense. Did we do that?

No. We did not do that.

You could store the string in a session variable. Then you'd only have
to parse it once per session. Now, session storage generates opaque
state that is hard to debug, which is messy and evil, but it would work.

We did do the storing. Did we do the reading of the stored value?

No. We did not do that.

You could even store the string in the database. Then you'd only have to
parse it once per installation (or per user, if we stored it as a
setting). This would of course Break Everything, as we wouldn't be able
to change the setting on file, and you probably have to log in a first
time to notice that you don't like the date format, and it wouldn't make
that much of a difference, as it's IPC vs IO, but the database can cache
pretty well, so it would be a tad better.

We did do this storing, in the inefficient, per-user way. Did we do the
reading of the stored value?

No. We did not do that. We tried to read a *different* value, but we did
not read the page we actually stored.

This patch throws away the anti-caching, overhead-generating fluff, and
replaces it with the common sense variety.

This makes the listview archive page a helluvalot quicker to generate,
as we no longer parse nagios.cfg (after running a couple of SQL queries)
twice per returned log row.

Change-Id: I7e97ca79ab72ef34825c0728816330f3ad6203bb
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>